Custom Grappelli used by SBGrid.
These are files that exist on our production server that have changes
that no SBGrid staff are aware of. These must have been pushed or
changed by our collobrators. Adding these to a git repo so that we
successfully redeploy our coe to new VM's. At some point someone is
going to have to clean this mess up.